Pcb: digital logic needs to run on 3.3V
The osicallator only works with 3.3V
(which I realized one year after designing this...)
thus we will run the whole logic with 3.3V
which works for all chips except for the
atmega which is now a little bit overclocked.
It still seems to work, but is not recommended
for production.